add comfirm, fix validate date...

+++ Frontend/src/app/Project/components/project-list/project-list.component.ts
@@ -1,24 +1,25 @@
-import { Component, ChangeDetectionStrategy, OnInit, OnDestroy, ChangeDetectorRef, NgModule, ViewChild, OnChanges, DoCheck, Output, EventEmitter } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit, OnDestroy, ChangeDetectorRef, ViewChild, Output, EventEmitter } from '@angular/core';
 import { Project, Status } from '../../models/project.model'
 import { ProjectServices } from '../../services/index'
-
 import { FormGroup, NgForm } from '@angular/forms';
-import { Observable } from 'rxjs';
-import { ProjectCreateDto } from 'src/app/swagger/models';
+
 import { Router } from '@angular/router';
 import { EmployeeServices } from 'src/app/Employee/services/employee.service';
-import { ngModuleJitUrl } from '@angular/compiler';
-import { multicast } from 'rxjs/operators';
+
+import { MessageService, ConfirmEventType, ConfirmationService } from 'primeng/api';
 
 @Component({
     selector: 'pim-project-list',
     styleUrls: ['./project-list.component.scss'],
     templateUrl: './project-list.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    providers: [ConfirmationService, MessageService]
 })
 export class ProjectListComponent implements OnInit {
     keysearch = "";
-    status = "";
+    status = { index: null, name: "" };
+    columnSort = "";
+    orderSort = 1;
     statusEnum = Status;
     nameChoose = "Projects List";
     listProject: Project[] = [];
@@ -33,7 +34,9 @@ export class ProjectListComponent implements OnInit {
     constructor(public projectServices: ProjectServices,
         private employeeServices: EmployeeServices,
         private cdr: ChangeDetectorRef,
-        private router: Router) {
+        private router: Router,
+        private confirmationService: ConfirmationService,
+        private messageService: MessageService) {
     }
 
     ngOnInit() {
@@ -55,8 +58,7 @@ export class ProjectListComponent implements OnInit {
     }
 
     onSubmit(searchForm) {
-        console.log(searchForm.value)
-        this.projectServices.getHaveCondition(searchForm.value.keysearch, searchForm.value.status, 1).subscribe(data => {
+        this.projectServices.getHaveCondition(searchForm.value.keysearch, searchForm.value.status.name, 1, this.columnSort, this.orderSort).subscribe(data => {
             this.listProject = data;
             this.projectServices.countProjects(this.keysearch, this.status).subscribe(count => {
                 this.totalRecords = count;
@@ -67,31 +69,55 @@ export class ProjectListComponent implements OnInit {
 
     }
 
-
-
-    onDeleteProject(id) {
-        this.projectServices.getDetailProject(id).subscribe(data => {
-            if (data.Status == 0) {
-                this.projectServices.deleteProject([id]).subscribe(() => {
-                    this.getAllProject();
-                });
+    onDeleteProject(id: number) {
+        this.confirmationService.confirm({
+            message: 'Are you sure that you want to delete?',
+            header: 'Confirmation',
+            icon: 'pi pi-info-circle',
+            accept: () => {
+                this.projectServices.getDetailProject(id).subscribe(data => {
+                    if (data.Status == 0) {
+                        this.projectServices.deleteProject([id]).subscribe(() => {
+                            this.getAllProject();
+                        });
+                        this.messageService.add({ severity: 'success', summary: 'Confirmed', detail: 'Project is deleted' });
+                    }
+                    else {
+                        this.messageService.add({ severity: 'warn', summary: 'Cancelled', detail: 'Delete failed' });
+                    }
+                })
             }
-            else {
-                console.log("shouldn't delete");
-            }
-        })
+        });
     }
     onDeleteProjects() {
-        let checkedIds = this.selectedProjects.map(x => x.Id);
-        this.projectServices.deleteProject(checkedIds).subscribe(() => {
-            this.getAllProject();
+        let total = this.selectedProjects.length;
+        this.confirmationService.confirm({
+            message: 'Are you sure that you want to delete ' + total + ' projects?',
+            header: 'Confirmation',
+            icon: 'pi pi-info-circle',
+            accept: () => {
+                let checkDelete = this.selectedProjects.filter(x => x.Status != 0);
+                if (checkDelete.length == 0) {
+                    let checkedIds = this.selectedProjects.map(x => x.Id);
+                    this.projectServices.deleteProject(checkedIds).subscribe(() => {
+                        this.getAllProject();
+                    });
+                    this.selectedProjects = [];
+                    this.messageService.add({ severity: 'success', summary: 'Confirmed', detail: total + ' projects are deleted' });
+                }
+                else {
+                    this.messageService.add({ severity: 'warn', summary: 'Cancelled', detail: 'Delete failed' });
+                }
+            }
         });
-        this.selectedProjects = [];
+
     }
 
     loadProjects(event) {
         let page = event.first / event.rows + 1;
-        this.projectServices.getHaveCondition(this.keysearch, this.status, page).subscribe(data => {
+        this.orderSort = event.sortOrder;
+        this.columnSort = event.sortField
+        this.projectServices.getHaveCondition(this.keysearch, this.status.name, page, this.columnSort, this.orderSort).subscribe(data => {
             this.listProject = data;
             this.projectServices.countProjects(this.keysearch, this.status).subscribe(count => {
                 this.totalRecords = count;
@@ -99,6 +125,5 @@ export class ProjectListComponent implements OnInit {
             });
             this.cdr.markForCheck();
         });
-
     }
 }
